move classes and methods from base.py to serialize.py
